+++ test/tree.spec.ts
@@ -1,5 +1,6 @@
 import { Tree } from '../src/tree';
-import { TreeModel, TreeModelSettings, FoldingType, CssClasses, ChildrenLoadingFunction } from '../src/tree.types';
+import { FoldingType, TreeModel, TreeModelSettings } from '../src/tree.types';
+import { NodeMenuItemAction } from '../src/menu/menu.events';
 
 describe('Tree', () => {
   it('should detect empty string', () => {
@@ -383,6 +384,10 @@ describe('Tree', () => {
 
     expect(servantTree.hasChild(child)).toEqual(true);
     expect(child.value).toEqual('');
+
+    expect(child.id).toBeDefined();
+    expect(child.id).toEqual(jasmine.any(String));
+
     expect(child.children).toEqual(null);
     expect(child.isLeaf()).toEqual(true);
     expect(child.isNew()).toEqual(true);
@@ -1125,4 +1130,88 @@ describe('Tree', () => {
     expect(tree.isBranch()).toBeTruthy();
   });
 
+  it('can be converted to TreeModel', () => {
+
+    const model: TreeModel = {
+      id: 6,
+      value: 'root',
+      emitLoadNextLevel: false,
+      settings: {
+        isCollapsedOnInit: true,
+        static: false,
+        leftMenu: false,
+        rightMenu: true
+      },
+      children: [
+        {
+          value: 'child#1',
+          emitLoadNextLevel: false,
+          settings: { isCollapsedOnInit: true, static: false, leftMenu: false, rightMenu: true } }
+      ]
+    };
+
+    const tree: Tree = new Tree(model);
+
+    expect(tree.toTreeModel()).toEqual(model);
+  });
+
+  it('has an access to menu items', () => {
+
+    const model: TreeModel = {
+      id: 42,
+      value: 'root',
+      settings: {
+        menuItems: [
+          {
+            action: NodeMenuItemAction.Custom,
+            name: 'FooMenuItem',
+            cssClass: 'fooMenuItemCss'
+          }
+        ]
+      }
+    };
+
+    const tree: Tree = new Tree(model);
+
+    expect(tree.hasCustomMenu()).toBe(true);
+    expect(tree.menuItems).toEqual([{
+      action: NodeMenuItemAction.Custom,
+      name: 'FooMenuItem',
+      cssClass: 'fooMenuItemCss'
+    }]);
+  });
+
+  it('static nodes cannot have custom menu', () => {
+
+    const model: TreeModel = {
+      id: 42,
+      value: 'root',
+      settings: {
+        static: true,
+        menuItems: [
+          {
+            action: NodeMenuItemAction.Custom,
+            name: 'FooMenuItem',
+            cssClass: 'fooMenuItemCss'
+          }
+        ]
+      }
+    };
+
+    const tree: Tree = new Tree(model);
+
+    expect(tree.hasCustomMenu()).toBe(false);
+  });
+
+  it('does not have custom menu without menu items', () => {
+
+    const model: TreeModel = {
+      id: 42,
+      value: 'root'
+    };
+
+    const tree: Tree = new Tree(model);
+
+    expect(tree.hasCustomMenu()).toBe(false);
+  });
 });